created components folder, moved homepage functions to components, installed react router for routing between pages
